feat: prefetch vertex images for anthropic models
